==> 2_DB2CSV.ipynb <==
import sqlite3 as sq
import pandas as pd

con = sq.connect("data/NPOPlayer.db")
def run_query(con, table, query):
    cur = con.cursor()
    cur.execute(query)
    results = cur.fetchall()
    return results

def convert_db_table_to_DF(con, table):
    header_query = "SELECT name FROM pragma_table_info('" + table + "') ORDER BY cid;"
    cols_init = run_query(con, table, header_query)
    cols = [cols_init[i][0] for i in range(len(cols_init))]
    content_query = "SELECT * from " + table
    data = run_query(con, table, content_query)
    df = pd.DataFrame(data, columns= cols)
    return df

df = convert_db_table_to_DF(con, "NPOPLAYER")

df.to_csv('data/DB.csv', sep=";")
