==> 3_PreComputeLDA.ipynb <==
#topic modeling with lda
import pandas as pd
from tqdm import tqdm
tqdm.pandas()

npo = pd.read_csv("data/DB.csv", sep=";")

npo[npo["longSummary"].isna()] = ""

npo['thumbnail']

description = npo['longSummary']

print(description)

#create function to clean the text
import spacy
import nltk
from nltk.tokenize import word_tokenize 
from nltk.corpus import stopwords
#nlp = spacy.load("nl_core_news_sm")
nltk.download('stopwords')
nltk.download('wordnet')
sw = set(stopwords.words('dutch'))

import re
import string
from nltk.stem import WordNetLemmatizer
lemmatizer = WordNetLemmatizer()

def clean_text(x):
    # remove punctutation
    x = x.translate(str.maketrans('','',string.punctuation))
    #remove digits
    x = x.translate(str.maketrans('','',string.digits))
    # tokenize for stopwords
    x = word_tokenize(x.lower())
    x = [lemmatizer.lemmatize(token) for token in x if token not in sw]
    x = ' '.join([w for w in x])
    return x

description_cleaned = description.progress_apply(clean_text)

print(description_cleaned)

import sklearn
from sklearn.feature_extraction.text import TfidfVectorizer

vectorizer = TfidfVectorizer(min_df=25)
tf_idf = vectorizer.fit_transform(description_cleaned)
columns=vectorizer.get_feature_names_out()

from sklearn.decomposition import LatentDirichletAllocation
n_topics = 15
lda_npo = LatentDirichletAllocation(n_components=n_topics, learning_method='online', n_jobs=16, verbose=1)
lda_npo.fit(tf_idf)

terms = columns
def df_lda(model,terms,n_topics):
    df = pd.DataFrame(model.components_, columns=terms)
    df = df.T
    print(df)
    for topic in range(n_topics):
        df_topic = df[topic]
        print('topic',topic,'top words are:\n',df[topic].sort_values(ascending=False)[:15])

print(df_lda(lda_npo, terms,n_topics))

# topics
"""
1: culture
2: 
3: 
4: children programming
5: 
6: 
7:
8:
9: 
10: 
"""


# find the most likely topic for each instance in the dataframe
topic_dist = lda_npo.transform(tf_idf)

clusters = topic_dist.argmax(axis=1)

data = {
  'description': description_cleaned,
  'topic': clusters
}
cluster_df = pd.DataFrame(data)

cluster_df['subtopic'] = None
print(cluster_df)

data_df = {
    'topic': [-1,-1],
    'subtopic': [-1,-1],
    'words': [['letter','car'],['letter','car']]
}
words_subtopic = pd.DataFrame(data_df)

def sub_topic(cluster_df,max_sub_topic_amount):
    for i in range(n_topics):
        # get the right instances
        cluster = cluster_df[cluster_df['topic']==i]
        sub_topic_amount = min(len(cluster)/20,max_sub_topic_amount)
        description = cluster['description']
        # create tf_idf
        vectorizer = TfidfVectorizer(min_df=10)
        tf_idf = vectorizer.fit_transform(description)
        columns=vectorizer.get_feature_names_out()
        # make lda and the clusters
        lda = LatentDirichletAllocation(n_components=sub_topic_amount, learning_method='online', n_jobs=16, verbose=1)
        lda.fit(tf_idf)
        df_subtopic = pd.DataFrame(lda.components_, columns=columns)
        df_subtopic = df_subtopic.T
        for topic in range(sub_topic_amount):
            df_topic = df_subtopic[topic]
            most_words = df_topic.sort_values(ascending=False)[:10].index.tolist()
            entry = [i,topic,most_words]
            words_subtopic.loc[len(words_subtopic)] = entry
        topic_dist = lda.transform(tf_idf)
        clusters = topic_dist.argmax(axis=1)
        #input the cluster to the right topic cluster
        cluster_df['subtopic'][cluster_df['topic']==i] = clusters
        

sub_topic(cluster_df,10)

print(words_subtopic)

words_subtopic.to_csv('topic_words.csv')

print(cluster_df)

sum(cluster_df['subtopic'].isna())

npo['topic'] = cluster_df['topic']

npo['subtopic'] = cluster_df['subtopic']

npo.to_csv('NPO_topics.csv')
