==> 4_DropSparseTopics.ipynb <==
import pandas as pd
from tqdm import tqdm

df_topics = pd.read_csv("data/NPO_topics.csv").drop(columns=['Unnamed: 0.1', 'Unnamed: 0',"icons"]).dropna()
df_topics['Topic'] = (df_topics['topic'].astype(str) + df_topics['subtopic'].astype(str)).astype(int)
df_topics.drop(columns=['topic', 'subtopic', 'kids'], inplace=True)

df_seasons = pd.read_csv("data/NPO_mike.csv")
df_seasons = df_seasons[["mediaID", "season"]]
df_result = pd.merge(df_topics, df_seasons).dropna().sample(5000)


df_topics[df_topics["mediaID"] == "POW_04415998"]

keep = dict(df_result['Topic'].value_counts() > 10)
df_result['keep'] = df_result["Topic"].apply(lambda x: keep[x])

df_result = df_result[df_result['keep']].drop(columns=['keep'])

df_result.to_csv("data/NPOPlayerv2.csv", index= False)
df_result.to_csv("Streamlit/data/NPOPlayerv2.csv", index= False)
