==> 5_PreComputeSearchResults.ipynb <==
import pandas as pd
from tqdm import tqdm
tqdm.pandas()
import json

df = pd.read_csv("streamlit/data/NPOPlayerv2.csv")

mediaID_MainTitleTokens = {}
mediaID_summaryTokens = {}

for index, row in tqdm(df.iterrows()):
    mediaID_MainTitleTokens[row["mediaID"]] = str(row["mainTitle"]).lower().split(" ")
    mediaID_summaryTokens[row["mediaID"]] = list(set((str(row["longSummary"])+ " " + str(row["shortSummary"])).lower().split(" ")))
#mediaID_MainTitleTokens
len(mediaID_summaryTokens)

mainTitle_mainTileTokens = {}
for mainTile in df["mainTitle"].unique():
    mainTitle_mainTileTokens[mainTile] = mainTile.lower().split(" ")

def save_activities():
    with open('streamlit/data/mainTitle_mainTitleTokens.json', 'w') as outfile:
        json.dump(mainTitle_mainTileTokens, outfile)
save_activities()


def save_activities():
    with open('streamlit/data/mediaID_MainTitleTokens.json', 'w') as outfile:
        json.dump(mediaID_MainTitleTokens, outfile)
save_activities()


""" DEPRECIATED
def save_activities():
    with open('streamlit/data/mediaID_summaryTokens.json', 'w') as outfile:
        json.dump(mediaID_summaryTokens, outfile)
save_activities()
"""
