==> 6_MergeFakeUserData.ipynb <==
File to merge al the generated fake user data (merging json files)

import json 
import pandas as pd

df1 = pd.read_json("data/FakeUserData/activitiesFemke.json")
df2 = pd.read_json("data/FakeUserData/activitiesLena.json")
df3 = pd.read_json("data/FakeUserData/activitiesMaarten.json")
df4 = pd.read_json("data/FakeUserData/activitiesMariaenZak.json")
df5 = pd.read_json("data/FakeUserData/activitiesHarm.json")

FakeUserData = pd.concat([df1,df2,df3,df4,df5], axis=0, ignore_index=True)

FakeUserData.to_dict(orient= 'records')
FakeUserData.to_json("Streamlit/data/activities.json", orient='records')
